+++ CHANGELOG.md
@@ -3,6 +3,10 @@ SPDX-FileCopyrightText: 2020 Robin Krahl <robin.krahl@ireas.org>
 SPDX-License-Identifier: CC0-1.0
 -->
 
+# Unreleased
+
+- Implement `From<&String>` for `StyledString`.
+
 # v0.1.0 (2020-10-15)
 
 Initial release with support for formatted text, text wrapping and basic